+++ src/main/java/be/baur/sds/NodeType.java
@@ -5,8 +5,8 @@ import java.util.List;
 import be.baur.sda.Node;
 import be.baur.sda.SDA;
 import be.baur.sda.DataNode;
-import be.baur.sds.content.AbstractStringType;
-import be.baur.sds.content.RangedType;
+import be.baur.sds.content.CharacterType;
+import be.baur.sds.content.ComparableType;
 import be.baur.sds.model.ModelGroup;
 import be.baur.sds.serialization.Attribute;
 import be.baur.sds.serialization.Components;
@@ -132,18 +132,18 @@ public class NodeType extends Type {
 		// facets are rendered ONLY if we are not a type reference!
 		if (getGlobalType() == null) {
 
-			final boolean stringType = (this instanceof AbstractStringType);
+			final boolean isCharType = (this instanceof CharacterType);
 
-			if (stringType) {
-				AbstractStringType<?> t = (AbstractStringType<?>) this;
+			if (isCharType) {
+				CharacterType<?> t = (CharacterType<?>) this;
 				if (t.getLength().min != 0 || t.getLength().max != Integer.MAX_VALUE)
 					node.add(new DataNode(Attribute.LENGTH.tag, t.getLength().toString()));
 			}
 	
-			if (this instanceof RangedType) {
-				RangedType<?> t = (RangedType<?>) this;
-				if (t.getRange().min != null || t.getRange().max != null)
-					node.add(new DataNode(Attribute.VALUE.tag, t.getRange().toString()));
+			if (this instanceof ComparableType) {
+				ComparableType<?> t = (ComparableType<?>) this;
+				if (t.getInterval().min != null || t.getInterval().max != null)
+					node.add(new DataNode(Attribute.VALUE.tag, t.getInterval().toString()));
 			}
 			
 			if (this instanceof DataType) {
@@ -152,7 +152,7 @@ public class NodeType extends Type {
 				if (t.getPattern() != null)
 					node.add(new DataNode(Attribute.PATTERN.tag, t.getPattern().toString()));
 				
-				if (stringType == !t.isNullable())
+				if (isCharType == !t.isNullable())
 					node.add(new DataNode(Attribute.NULLABLE.tag, String.valueOf(t.isNullable())));
 			}
 		}